Add node-rdkafka as a dependency
